+++ CHANGELOG.md
@@ -99,6 +99,7 @@ The list of direct contributors to the MAME 2003-Plus github repository can be f
 * Monster Maulers
 * Mouse Shooter GoGo
 * Ms. Pacman Twin
+* Mrs. Dynamite
 * Multi Champ Deluxe * Mature *
 * Oo Parts (Japan, Prototype)
 * Opa Opa (Rev A Unprotected)
@@ -140,6 +141,7 @@ The list of direct contributors to the MAME 2003-Plus github repository can be f
 * Sonic Boom
 * Space Fever High Splitter
 * Space Launcher
+* Space Raider
 * Squash
 * Star Fire 2
 * Steel Worker
@@ -709,3 +711,4 @@ Gun & Fronter, Metal Black, Ninja Kids, and Pulirula [arcadez]
 * Updated outrun.c to MAME89 and then fixed the sound and graphical road drawing plus hooked up the inputs so that Turbo Out Run can now be played in this core [arcadez]
 * Added support for rare, Ms. Pacman Twin (mspactwin) simultaneous 2 player play [arcadez, grant2258, mahoneyt944]
 * Hooked up HD647180 MCU sound support for Ghox removing the need for hacked samples plus added a dedicated joystick version of the game [grant2258, arcadez]
+* Added support for Mrs. Dynamite and Space Raider to ladybug.c also updated the SN76496 sound core to allow for 5 chips to be used for both of these games [arcadez]